+++ R/readAudio.R
@@ -8,91 +8,99 @@
 #' @param mime MIME type of file to read, or "auto". Supported types are "audio/x-wav" and "audio/mpeg" (MP3)
 #' @param from Start point in file to return
 #' @param to End point in file to return
-#' @param units One of "samples", "seconds", "minutes", "hours"
+#' @param units One of "samples", "seconds", "minutes", "hours". Default is "seconds".
 #' @return A Wave object
 #' @export
-#' @importFrom tuneR readMP3 readWave Wave
+#' @importFrom tuneR readMP3 readWave stereo Wave
 #' @importFrom seewave cutw
 #' @importFrom tools file_ext
 #' @importFrom mime guess_type
-readAudio <- function(file, mime="auto", from=1, to=Inf, units="samples") {
+readAudio <- function(file, mime="auto", from=0, to=Inf, units="seconds") {
   if (mime == "auto") {
     mime <- guess_type(file)
   }
+  if (units=="samples" & from == 0) {
+    fromS <- 1
+  } else{
+    fromS <- from
+  }
 
   if (mime == "audio/x-wav") {
-    wave <- readWave(file, from=from, to=to, units=units)
-    return(wave)
+    tryCatch({
+      wave <- readWave(file, from=fromS, to=to, units=units)
+      return(wave)
+    },
+    error=function(cond){
+
+    })
   }
 
   if (mime=="audio/mpeg") {
-    #Check if av package available
-    if (package.installed("av", askInstall=FALSE)==FALSE) {
-
+    wave <- NULL
+    wave <- tryCatch({
       wave <- readMP3(file)
-      if (from==1 && to==Inf) {
+      if (units=="samples") {
+        return(cutws(wave, fromS, to))
+      }
+      if (from==0 && to==Inf) {
         return(wave)
       }
-
-      if (units=="samples") {
-        return(cutws(wave, from, to))
-      }
-
       from <- convert2seconds(from, input=units)
-      to <- convert2seconds(from, input=units)
-      return(cutw(wave,from=from, to=to))
+      to <- convert2seconds(to, input=units)
+      return(cutw(wave,from=from, to=to, output="Wave"))
+    },
+    error=function(cond){
+    })
+  }
+
+  #Check if av package available
+  if (package.installed("av", askInstall=TRUE)) {
+    #Using av package
+    channels <- av::av_media_info(file)$audio[['channels']]
+    if (is.null(channels)) {
+      stop("Could not determine number of channels.")
     }
-    #av available - continue
-  } else {
-    message("Unidentified MIME type - attempting to read using package av")
-    package.installed("av")
-  }
+    if (channels > 2) {
+      stop("channel count greater than 2 is not supported")
+    }
 
-  #Using av package
-  channels <- av::av_media_info(file)$audio[['channels']]
-  if (channels > 2) {
-    stop("channel count greater than 2 is not supported")
-  }
-
-  if (from==1 && to==Inf) {
     wave <- av::read_audio_bin(file, channels=channels)
-  } else {
-    if (units != "samples") {
-      from <- convert2seconds(from, input=units)
-      to <- convert2seconds(from, input=units)
-      wave <- av::read_audio_bin(file, channels=1, start_time=from, end_time=to)
+    wave[which(is.na(wave))] <- 0
+    bit <- .bitdepth(wave)
+
+    if (channels == 1) {
+      wave <- Wave(left=wave, samp.rate=attr(wave, "sample_rate"), bit=bit)
+    }
+    if (channels == 2) {
+      left <- wave[seq(1, length(wave), by = 2)]
+      right <- wave[seq(2, length(wave), by = 2)]
+      wave <- Wave(left=left, right=right, samp.rate=attr(wave, "sample_rate"), bit=bit)
+    }
+
+    if (units == "samples") {
+      return(cutws(wave,from=fromS, to=to))
+    } else {
+      if (from==0 & to == Inf) {
+        return(wave)
+      }
+      if (channels == 1) {
+        return(cutw(wave, from=convert2seconds(from, units), to=convert2seconds(to, units), output="Wave"))
+      }
+      if (channels == 2) {
+        left <- cutw(wave, channel=1, from=convert2seconds(from, units), to=convert2seconds(to, units), output="Wave")
+        right <- cutw(wave, channel=2, from=convert2seconds(from, units), to=convert2seconds(to, units), output="Wave")
+        return(stereo(left, right))
+      }
     }
   }
-
-  wave[which(is.na(wave))] <- 0
-  bit <- bitdepth(wave)
-
-  if (channels == 1) {
-    wave <- Wave(left=wave, samp.rate=attr(wave, "sample_rate"), bit=bit)
-    return(wave)
-  }
-  if (channels == 2) {
-    left <- wave[seq(1, length(wave), by = 2)]
-    right <- wave[seq(2, length(wave), by = 2)]
-    wave <- Wave(left=left, right=right, samp.rate=attr(wave, "sample_rate"), bit=bit)
-    d <- 2
-    return(wave)
-  }
-
-  if (units == "samples") {
-    return(cutws(wave,from=from, to=to))
-  } else {
-    return(wave)
-  }
-
   stop("File could not be processed")
 }
 
-bitdepth <- function(v) {
-  m <- max(abs(v), na.rm=TRUE)
+.bitdepth <- function(v) {
+  m <- ceiling(max(abs(v), na.rm=TRUE))
   if (m <= 128) { return(8) }
   if (m <= 32768) { return(16) }
+  if (m <= 8388608) { return(24) }
   if (m <= 2147483648) { return(32) }
-  if (m <= 9223372036854775808) { return(64) }
-  stop("Bit depths above 64bit are not supported")
+  stop("Bit depths above 32bit are not supported.")
 }