+++ R/typicalVolume.R
@@ -7,18 +7,22 @@
 #' @return Typical volume of thing in dBA, or if no thing parameter a data frame of all volumes
 #' @examples
 #' typicalVolume()
-#' typicalVolume("rocket")
+#' typicalVolume("rocket launch")
 #'
-typicalVolume <- function(thing="") {
-  tv <- typicalVolumes()
-  if (thing %in% tv[,1]) {
-    return(tv[tv$thing==thing,2])
+typicalVolume <- function(thing=NA_character_) {
+  tv <- .typicalVolumes()
+  if (is.na(thing)) {
+    return(tv)
   }
-  return(tv)
+  if (thing %in% tv[,1]) {
+    return(as.numeric(tv[tv$thing==thing,2]))
+  } else {
+    stop("Thing not found.")
+  }
 }
 
-typicalVolumes <- function() {
-  # Source: Murray Schaefer Soundscapes
+.typicalVolumes <- function() {
+  # Source: Murray Schafer Soundscapes
   n <- c(
     "steam engine",
     "printing works",