+++ R/windowing.R
@@ -1,42 +1,77 @@
 #' Windowing Function for Wave Objects
 #'
-#' Separates a Wave object into windows of a defined length and runs a function on the window section. Windows may overlap, and the function can make use of 'parallel' package for multicore processing.
+#' Separates a Wave object into windows of a defined length and runs a function
+#' on the window section. Windows may overlap, and the function can make use of
+#' 'parallel' package for multi-core processing. It will also show a progress bar
+#' if the 'pbapply' package is installed.
 #'
-#' @param wave A Wave object
-#' @param window.length The lag used to create the A-matrix
-#' @param window.overlap A matrix used to code the Duration-Shape pairs
-#' @param bind.wave If TRUE and FUN returns wave objects these are combined into a single object
-#' @param FUN If TRUE plots the workings of the coding algorithm
+#' @param wave A Wave object or filename. Using filenames may save loading an
+#'   entire large file into memory.
+#' @param window.length The length of the analysis window (in samples).
+#' @param window.overlap The overlap between successive windows (in samples), a
+#'   negative value will result in a gap between windows.
+#' @param bind.wave If TRUE and FUN returns wave objects, then these are
+#'   combined into a single object
+#' @param FUN FUN to be applied to windows.
 #' @param ... Additional parameters to FUN
-#' @param cluster A cluster form the 'parallel' package for multicore computation
+#' @param complete.windows If TRUE (default) the final window will not be
+#'   processed unless it has a length equal to window.length.
+#' @param cluster A cluster form the 'parallel' package for multi-core computation.
 #' @keywords wave
 #' @export
 #' @examples
 #' \dontrun{
-#' windowing(wave, window.length=1000, window.overlap=0, bind.wave=TRUE, FUN=noChange)
+#' windowing(wave, window.length=1000, FUN=duration, window.overlap=0, bind.wave=TRUE)
 #' }
-
 windowing <- function(
   wave,
   window.length,
+  FUN,
   window.overlap=0,
   bind.wave=TRUE,
-  FUN,
-  ...,
-  cluster=NULL){
-  n.windows <- ceiling(length(wave@left) / (window.length - window.overlap))
+  complete.windows=TRUE,
+  cluster=NULL,
+  ...){
+  if (typeof(wave) == "character") {
+    FUN2 <- function(start, wave, window.length, ...){
+      wave <- readAudio(wave, from=start,to=start+window.length,units="samples")
+      return(FUN(wave, start, window.length, ...))
+    }
+    info <- av::av_media_info(wave)
+    n.samples <- info$duration * info$audio[,"sample_rate"]
+  } else {
+    validateIsWave(wave)
+    n.samples <- length(wave)
+    FUN2 <- function(start, wave, window.length, ...){
+      section <- cutws(wave, from=start, to=start+window.length)
+      return(FUN(section, start, window.length, ...))
+    }
+  }
+  n.windows <- ceiling(n.samples / (window.length - window.overlap))
+
   starts <- c(
     1,
     1:(n.windows-1) * (window.length - window.overlap) +1
   )
-
-  if (is.null(cluster)){
-    l <- lapply(starts, FUN, wave=wave, window.length=window.length, ...)
-  } else {
-    l <- parLapply(cluster, starts, FUN, wave=wave, window.length=window.length, ...)
+  if (complete.windows) {
+    starts <- starts[which(starts <= n.samples - window.length + 1)]
   }
 
-  if (typeof(l[[1]]) == "S4" & class(l[[1]])[[1]] == "Wave") {
+  if (is.null(cluster)){
+    if (package.installed("pbapply", askInstall = FALSE)) {
+      l <- pbapply::pblapply(starts, FUN2, wave=wave, window.length=window.length, ...)
+    } else {
+      l <- lapply(starts, FUN2, wave=wave, window.length=window.length, ...)
+    }
+  } else {
+    if (package.installed("pbapply", askInstall = FALSE)) {
+      l <- pbapply::pblapply(starts, FUN2, cl=cluster, wave=wave, window.length=window.length, ...)
+    } else {
+      l <- parallel::parLapply(cluster, starts, FUN2, wave=wave, window.length=window.length, ...)
+    }
+  }
+
+  if (bind.wave & typeof(l[[1]]) == "S4" & class(l[[1]])[[1]] == "Wave") {
     w <- l[[1]]
     for (i in 2:length(l)) {
       w <- tuneR::bind(w, l[[i]])
@@ -46,24 +81,3 @@ windowing <- function(
   return(l)
 }
 
-#' List available windowing functions
-#'
-#' Lists all available windowing functions.
-#'
-#' @keywords wave
-#' @export
-#' @importFrom utils getFromNamespace
-#' @examples
-#' \dontrun{
-#' windowing.functions()
-#' }
-
-windowing.functions <- function() {
-  ip <- installed.packages()[,1]
-  for (i in 1:length(ip)) {
-    if (exists('ssRwindows', where=asNamespace(ip[i]))) {
-      l <- do.call(getFromNamespace("ssRwindows", ip[i]), args=list(), envir=asNamespace(ip[i]))
-      print(l)
-    }
-  }
-}