+++ tests/testthat/test-frequencyStats.R
@@ -1,8 +1,6 @@
-context("frequencyStats")
-
 test_that("inputs are correct", {
-  expect_error(frequencyStats("string"), "frequencyStats expects a Wave object")
-  expect_error(frequencyStats(1), "frequencyStats expects a Wave object")
+  expect_error(frequencyStats("string"), "Expecting a Wave object")
+  expect_error(frequencyStats(1), "Expecting a Wave object")
 })
 
 test_that("plotting is ok", {